scripts/requirements.txt
fabric
pyyaml

invoke